+++ examples/BasicOperations/AddCampaigns.php
@@ -51,7 +51,7 @@ class AddCampaigns
     {
         // Either pass the required parameters for this example on the command line, or insert them
         // into the constants above.
-        $options = ArgumentParser::parseCommandArguments([
+        $options = (new ArgumentParser())->parseCommandArguments([
             ArgumentNames::CUSTOMER_ID => GetOpt::REQUIRED_ARGUMENT
         ]);