+++ examples/BasicOperations/AddKeywords.php
@@ -48,7 +48,7 @@ class AddKeywords
     {
         // Either pass the required parameters for this example on the command line, or insert them
         // into the constants above.
-        $options = ArgumentParser::parseCommandArguments([
+        $options = (new ArgumentParser())->parseCommandArguments([
             ArgumentNames::CUSTOMER_ID => GetOpt::REQUIRED_ARGUMENT,
             ArgumentNames::AD_GROUP_ID => GetOpt::REQUIRED_ARGUMENT,
             ArgumentNames::KEYWORD_TEXT => GetOpt::OPTIONAL_ARGUMENT