+++ examples/BasicOperations/UpdateKeyword.php
@@ -50,7 +50,7 @@ class UpdateKeyword
     {
         // Either pass the required parameters for this example on the command line, or insert them
         // into the constants above.
-        $options = ArgumentParser::parseCommandArguments([
+        $options = (new ArgumentParser())->parseCommandArguments([
             ArgumentNames::CUSTOMER_ID => GetOpt::REQUIRED_ARGUMENT,
             ArgumentNames::AD_GROUP_ID => GetOpt::REQUIRED_ARGUMENT,
             ArgumentNames::CRITERION_ID => GetOpt::REQUIRED_ARGUMENT