+++ src/Google/Ads/GoogleAds/V0/Services/Gapic/KeywordViewServiceGapicClient.php
@@ -115,7 +115,6 @@ class KeywordViewServiceGapicClient
                     'restClientConfigPath' => __DIR__.'/../resources/keyword_view_service_rest_client_config.php',
                 ],
             ],
-            'versionFile' => __DIR__.'/../../../VERSION',
         ];
     }
 
@@ -242,8 +241,8 @@ class KeywordViewServiceGapicClient
      *               'grpc' => [...],
      *               'rest' => [...]
      *           ];
-     *           See the {@see \Google\Cloud\ApiCore\Transport\GrpcTransport::build()} and
-     *           {@see \Google\Cloud\ApiCore\Transport\RestTransport::build()} methods for the
+     *           See the {@see \Google\ApiCore\Transport\GrpcTransport::build()} and
+     *           {@see \Google\ApiCore\Transport\RestTransport::build()} methods for the
      *           supported options.
      * }
      *