+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+*   Added support for V0_2 of Google Ads API, which includes the Percent CPC
+    bidding strategy.
+
 ## 0.1.0
 
 *   Initial release with support for V0_1 of Google Ads API.